+++ example/lib/view_only.dart
@@ -1,4 +1,3 @@
-
 import 'package:example/selection_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_muscle_anatomy/body/body.dart';
@@ -6,6 +5,8 @@ import 'package:flutter_muscle_anatomy/core/core.dart';
 import 'package:flutter_svg/svg.dart';
 
 enum Gender { male, female }
+
+const Size viewSize = Size(200, 300);
 
 class ViewOnly extends StatefulWidget {
   const ViewOnly({super.key});
@@ -53,68 +54,19 @@ class _ViewOnlyState extends State<ViewOnly> {
       mainAxisAlignment: MainAxisAlignment.center,
       crossAxisAlignment: CrossAxisAlignment.center,
       children: [
-        SvgPicture.string(anatomy.toString(), width: 200, height: 300),
+        SvgPicture.string(
+          anatomy.toString(),
+          width: viewSize.width,
+          height: viewSize.height,
+        ),
         Text('SVG: $name'),
         const SizedBox(height: 16),
-        SizedBox(
-          width: 200,
-          height: 300,
-          child: CustomPaint(painter: AnatomyPainter(anatomy: anatomy)),
+        SizedBox.fromSize(
+          size: viewSize,
+          child: CustomPaint(painter: anatomy.customPainter(viewSize)),
         ),
         Text('Path Drawing: $name'),
       ],
     );
-  }
-}
-
-/// A custom painter that demonstrates how to use raw Path objects
-/// from the MuscleAnatomy model for custom rendering.
-class AnatomyPainter extends CustomPainter {
-  final MuscleAnatomy anatomy;
-
-  AnatomyPainter({required this.anatomy});
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    // 1. Calculate scale and offset to center the anatomy model into the available size
-    final scaledSize = anatomy.scaledSize(size);
-    final scale = scaledSize.width / anatomy.dimension.width;
-
-    final dx = (size.width - scaledSize.width) / 2;
-    final dy = (size.height - scaledSize.height) / 2;
-
-
-    final matrix = Matrix4.identity()
-      ..translateByDouble(dx, dy,0,1)
-      ..scaleByDouble(scale,scale, scale, 1.0);
-
-    // 2. Draw outline
-    for (final path in anatomy.outlinePaths) {
-      canvas.drawPath(path.transform(matrix.storage), anatomy.outlinePaint);
-    }
-
-    // 3. Draw all muscles
-    for (final muscleMember in anatomy.getMuscleMembers()) {
-      for (final path in muscleMember.paths) {
-        canvas.drawPath(
-          path.transform(matrix.storage),
-          muscleMember.decoration.strokePaint(),
-        );
-        canvas.drawPath(
-          path.transform(matrix.storage),
-          muscleMember.decoration.fillPaint(),
-        );
-      }
-    }
-
-    // 4. Draw hair
-    for (final path in anatomy.hairOutlinePaths) {
-      canvas.drawPath(path.transform(matrix.storage), anatomy.hairFillPaint);
-    }
-  }
-
-  @override
-  bool shouldRepaint(AnatomyPainter oldDelegate) {
-    return oldDelegate.anatomy != anatomy;
   }
 }