+++ lib/flutter_muscle_anatomy.dart
@@ -4,7 +4,14 @@
 /// and allows for precise muscle highlighting.
 library;
 
-export 'body/body.dart' show Male, Female, MuscleAnatomy, Anatomy;
+export 'body/body.dart'
+    show
+        Male,
+        Female,
+        MuscleAnatomy,
+        Anatomy,
+        ViewScale,
+        MuscleInteractiveView;
 export 'core/core.dart'
     show
         Muscle,